+++ CHANGELOG.rst
@@ -5,7 +5,8 @@ Changelog
 =========
 
 * :feature:`13` Add a `.WCSCards` macro that expands into full WCS header information.
-* :feature:`-` Allow MacroCard.macro to return `.Card` or `.CardGroup`.
+* :feature:`-` Allow `.MacroCard.macro` to return `.Card` or `.CardGroup`.
+* :feature:`-` Allow `.HeaderModel` to accept ``None`` as an item. This is useful to programmatically define cards that in some cases may not be added.
 
 * :release:`0.4.0 <2021-02-12>`
 * :feature:`-` `.Card` now accepts a ``type`` to which to be cast. By default ``autocast=True`` will try to cast the value to the correct type after evaluating.